+++ src/read-sound-file.h
@@ -39,10 +39,10 @@ unsigned ca_sound_file_get_nchannels(ca_sound_file *f);
 unsigned ca_sound_file_get_rate(ca_sound_file *f);
 ca_sample_type_t ca_sound_file_get_sample_type(ca_sound_file *f);
 
-size_t ca_sound_file_get_size(ca_sound_file *f);
+off_t ca_sound_file_get_size(ca_sound_file *f);
 
-int ca_sound_file_read_int16(ca_sound_file *f, int16_t *d, unsigned *n);
-int ca_sound_file_read_uint8(ca_sound_file *f, uint8_t *d, unsigned *n);
+int ca_sound_file_read_int16(ca_sound_file *f, int16_t *d, size_t *n);
+int ca_sound_file_read_uint8(ca_sound_file *f, uint8_t *d, size_t *n);
 
 int ca_sound_file_read_arbitrary(ca_sound_file *f, void *d, size_t *n);
